fix(deadline): display deadline as datestring
